requirements.txt
streamlit==1.36.0
PyMuPDF==1.23.9
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0